# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/shopProducts": {
            "get": {
                "tags": [
                    "shop_product"
                ],
                "summary": "Get a listing of the shop_products.",
                "description": "Get all shop_products",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/shop_product"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "shop_product"
                ],
                "summary": "Store a newly created shop_product in storage",
                "description": "Store shop_product",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "shop_product that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/shop_product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/shop_product"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/shopProducts/{id}": {
            "get": {
                "tags": [
                    "shop_product"
                ],
                "summary": "Display the specified shop_product",
                "description": "Get shop_product",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of shop_product",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/shop_product"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "shop_product"
                ],
                "summary": "Update the specified shop_product in storage",
                "description": "Update shop_product",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of shop_product",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "shop_product that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/shop_product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/shop_product"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "shop_product"
                ],
                "summary": "Remove the specified shop_product from storage",
                "description": "Delete shop_product",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of shop_product",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "shop_product": {
            "required": [
                ""
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "sort": {
                    "description": "\u0421\u043e\u0440\u0442\u0438\u0440\u043e\u0432\u043a\u0430",
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "description": "\u0423\u043d\u0438\u043a\u0430\u043b\u044c\u043d\u043e\u0435 \u0438\u043c\u044f",
                    "type": "string"
                },
                "name_lang": {
                    "description": "\u041f\u0435\u0440\u0435\u0432\u043e\u0434\u044b \u0423\u043d\u0438\u043a\u0430\u043b\u044c\u043d\u043e\u0435 \u0438\u043c\u044f",
                    "type": "string"
                },
                "title": {
                    "description": "\u041d\u0430\u0437\u0432\u0430\u043d\u0438\u0435",
                    "type": "string"
                },
                "title_lang": {
                    "description": "\u041f\u0435\u0440\u0435\u0432\u043e\u0434\u044b \u041d\u0430\u0437\u0432\u0430\u043d\u0438\u0435",
                    "type": "string"
                },
                "tranz": {
                    "description": "\u0422\u0440\u0430\u043d\u0437\u0430\u043a\u0446\u0438\u044f",
                    "type": "boolean"
                },
                "accept": {
                    "description": "\u041f\u0440\u0438\u043d\u044f\u0442\u043e",
                    "type": "boolean"
                },
                "active": {
                    "description": "\u0410\u043a\u0442\u0438\u0432\u043d\u044b\u0439",
                    "type": "boolean"
                },
                "user_company_id": {
                    "description": "\u041a\u043e\u043c\u043f\u0430\u043d\u0438\u044f",
                    "type": "integer",
                    "format": "int32"
                },
                "shop_category_id": {
                    "description": "\u041a\u0430\u0442\u0435\u0433\u043e\u0440\u0438\u044f",
                    "type": "integer",
                    "format": "int32"
                },
                "package": {
                    "description": "\u0423\u043f\u0430\u043a\u043e\u0432\u043a\u0430",
                    "type": "string"
                },
                "shop_option": {
                    "description": "\u041f\u0430\u0440\u0430\u043c\u0435\u0442\u0440\u044b \u043e\u043f\u0446\u0438\u0439",
                    "type": "string"
                },
                "text": {
                    "description": "\u041f\u043e\u043b\u043d\u043e\u0435 \u043e\u043f\u0438\u0441\u0430\u043d\u0438\u0435 \u043f\u0440\u043e\u0434\u0443\u043a\u0442\u0430",
                    "type": "string"
                },
                "text_lang": {
                    "description": "\u041f\u0435\u0440\u0435\u0432\u043e\u0434\u044b \u041f\u043e\u043b\u043d\u043e\u0435 \u043e\u043f\u0438\u0441\u0430\u043d\u0438\u0435 \u043f\u0440\u043e\u0434\u0443\u043a\u0442\u0430",
                    "type": "string"
                },
                "image": {
                    "description": "\u0418\u0437\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u0438\u044f",
                    "type": "string"
                },
                "shop_option_ids": {
                    "description": "\u041e\u043f\u0446\u0438\u0438 \u043f\u0440\u043e\u0434\u0443\u043a\u0442\u0430",
                    "type": "string"
                },
                "shop_brand_id": {
                    "description": "\u0411\u0440\u0435\u043d\u0434",
                    "type": "integer",
                    "format": "int32"
                },
                "related": {
                    "description": "\u0421\u0432\u044f\u0437\u0430\u043d\u043d\u044b\u0435 \u0442\u043e\u0432\u0430\u0440\u044b",
                    "type": "string"
                },
                "shelf_life": {
                    "description": "\u0421\u0440\u043e\u043a \u0433\u043e\u0434\u043d\u043e\u0441\u0442\u0438",
                    "type": "integer",
                    "format": "int32"
                },
                "shelf_life_unit": {
                    "description": "\u0415\u0434\u0438\u043d\u0438\u0446\u0430 \u0438\u0437\u043c\u0435\u0440\u0435\u043d\u0438\u044f \u0441\u0440\u043e\u043a\u0430",
                    "type": "string"
                },
                "weight": {
                    "description": "\u0412\u0435\u0441 \u0432 \u043a\u0433",
                    "type": "number",
                    "format": "number"
                },
                "size": {
                    "description": "\u0420\u0430\u0437\u043c\u0435\u0440\u044b",
                    "type": "string"
                },
                "offer": {
                    "description": "\u0421\u043f\u0435\u0446\u0438\u0430\u043b\u044c\u043d\u044b\u0435 \u043f\u0440\u0435\u0434\u043b\u043e\u0436\u0435\u043d\u0438\u044f",
                    "type": "string"
                },
                "rating": {
                    "description": "\u0420\u0435\u0439\u0442\u0438\u043d\u0433",
                    "type": "number",
                    "format": "number"
                },
                "measure": {
                    "description": "\u0415\u0434\u0438\u043d\u0438\u0446\u0430 \u0438\u0437\u043c\u0435\u0440\u0435\u043d\u0438\u044f",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "\u0412\u0440\u0435\u043c\u044f \u0443\u0434\u0430\u043b\u0435\u043d\u0438\u044f",
                    "type": "string",
                    "format": "date-time"
                },
                "deleted_by": {
                    "description": "\u0423\u0434\u0430\u043b\u0451\u043d \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u0435\u043c",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_text": {
                    "description": "\u041a\u043e\u043c\u043c\u0435\u043d\u0442\u0430\u0440\u0438\u044f \u0443\u0434\u0430\u043b\u0435\u043d\u0438\u044f",
                    "type": "string"
                },
                "created_at": {
                    "description": "\u0412\u0440\u0435\u043c\u044f \u0441\u043e\u0437\u0434\u0430\u043d\u0438\u044f",
                    "type": "string",
                    "format": "date-time"
                },
                "modified_at": {
                    "description": "\u0412\u0440\u0435\u043c\u044f \u0438\u0437\u043c\u0435\u043d\u0435\u043d\u0438\u044f",
                    "type": "string",
                    "format": "date-time"
                },
                "created_by": {
                    "description": "\u0421\u043e\u0437\u0434\u0430\u043d \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u0435\u043c",
                    "type": "integer",
                    "format": "int32"
                },
                "modified_by": {
                    "description": "\u0418\u0437\u043c\u0435\u043d\u0435\u043d \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u0435\u043b\u0435\u043c",
                    "type": "integer",
                    "format": "int32"
                }
            }
        }
    }
}